requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==22.1.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
ciso8601==2.2.0
clang==5.0
click==8.0.4
colorama==0.4.4
cycler==0.11.0
cytoolz==0.11.2
decorator==4.4.2
Django==3.2.12
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
flatbuffers==1.12
frozenlist==1.2.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
hexbytes==0.2.2
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipfshttpclient==0.8.0a2
joblib==1.1.0
jsonschema==3.2.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lime==0.1.1.37
lru-dict==1.1.7
Markdown==3.3.6
matplotlib==3.3.4
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
networkx==2.5.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parsimonious==0.8.1
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
plotly==5.6.0
progressbar==2.5
protobuf==3.19.4
py-geth==3.7.0
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.1
pycryptodome==3.14.1
pyfiglet==0.8.post1
Pygments==2.11.2
pyparsing==3.0.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.8
schedule==1.1.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scs==3.2.0
semantic-version==2.9.0
shap==0.34.0
six==1.15.0
sqlparse==0.4.2
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
tomli==1.2.3
toolz==0.11.2
tqdm==4.63.0
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
webencodings==0.5.1
websockets==9.1
Werkzeug==2.0.3
wrapt==1.12.1
yarl==1.7.2
zipp==3.6.0